requirements.txt
absl-py==0.9.0
autograd==1.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
future==0.18.2
google-auth==1.20.1
google-auth-oauthlib==0.4.1
grpcio==1.31.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.0
networkx==2.4
numpy==1.19.1
oauthlib==3.1.0
Pillow==8.3.2
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
# tensorboard==2.3.0
# tensorboard-plugin-wit==1.7.0
tifffile==2020.7.24
# torch==1.6.0
# torchvision==0.7.0
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.1.0
